requirements.txt
alembic==1.16.2
blinker==1.9.0
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==46.0.2
distlib==0.3.9
dnspython==2.8.0
email-validator==2.0.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pdfkit==1.0.0
platformdirs==4.3.8
pycparser==2.23
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
virtualenv==20.31.2
Werkzeug==3.0.1
wkhtmltopdf==0.2
WTForms==3.2.1
reportlab==4.0.9
weasyprint==60.1
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-cov==4.1.0
python-json-logger==2.0.7
marshmallow==3.20.1
flask-caching==2.1.0
